+++ src/main/java/roomescape/domain/reservation/repository/MemoryReservationRepository.java
@@ -2,9 +2,9 @@ package roomescape.domain.reservation.repository;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.reservation.domain.model.Reservation;
-import roomescape.domain.reservation.domain.model.ReservationGuestName;
-import roomescape.domain.reservation.domain.model.ReservationTimeStamp;
+import roomescape.domain.reservation.model.Reservation;
+import roomescape.domain.reservation.model.ReservationGuestName;
+import roomescape.domain.reservation.model.ReservationTimeStamp;
 
 import java.util.ArrayList;
 import java.util.List;
