+++ src/main/java/roomescape/domain/reservation/service/ReservationQueryService.java
@@ -2,7 +2,7 @@ package roomescape.domain.reservation.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import roomescape.domain.reservation.domain.model.Reservation;
+import roomescape.domain.reservation.model.Reservation;
 import roomescape.domain.reservation.repository.ReservationRepository;
 import roomescape.domain.reservation.service.response.ReservationQueryResponse;
