+++ src/main/java/roomescape/domain/reservation/service/response/ReservationQueryResponse.java
@@ -2,7 +2,7 @@ package roomescape.domain.reservation.service.response;
 
 import lombok.Builder;
 import lombok.Getter;
-import roomescape.domain.reservation.domain.model.Reservation;
+import roomescape.domain.reservation.model.Reservation;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
