+++ src/test/java/roomescape/domain/reservation/domain/model/ReservationTest.java
@@ -2,6 +2,10 @@ package roomescape.domain.reservation.domain.model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.domain.reservation.model.Reservation;
+import roomescape.domain.reservation.model.ReservationGuestName;
+import roomescape.domain.reservation.model.ReservationStatus;
+import roomescape.domain.reservation.model.ReservationTimeStamp;
 import roomescape.global.infrastructure.ClockHolder;
 import roomescape.mock.FakeClockHolder;
