+++ src/test/java/roomescape/domain/reservation/service/ReservationQueryServiceTest.java
@@ -3,10 +3,10 @@ package roomescape.domain.reservation.service;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import roomescape.domain.reservation.domain.model.Reservation;
-import roomescape.domain.reservation.domain.model.ReservationGuestName;
-import roomescape.domain.reservation.domain.model.ReservationStatus;
-import roomescape.domain.reservation.domain.model.ReservationTimeStamp;
+import roomescape.domain.reservation.model.Reservation;
+import roomescape.domain.reservation.model.ReservationGuestName;
+import roomescape.domain.reservation.model.ReservationStatus;
+import roomescape.domain.reservation.model.ReservationTimeStamp;
 import roomescape.domain.reservation.repository.ReservationRepository;
 import roomescape.domain.reservation.service.response.ReservationQueryResponse;
 import roomescape.support.IntegrationTestSupport;
